+++ src/frykit/plot/api.py
@@ -33,8 +33,8 @@ from numpy.typing import ArrayLike, NDArray
 from shapely.geometry.base import BaseGeometry
 
 from frykit import get_data_dirpath
+from frykit._config import DataSource, config
 from frykit.calc import asarrays, get_values_between, lon_to_180
-from frykit.option import DataSource, get_option, validate_option
 from frykit.plot.artist import (
     Compass,
     Frame,
@@ -833,9 +833,9 @@ def label_cn_district(
 
 def _resolve_strict_clip(strict_clip: bool | None) -> bool:
     if strict_clip is None:
-        return get_option("strict_clip")
+        return config.strict_clip
     else:
-        validate_option("strict_clip", strict_clip)
+        config.validate("strict_clip", strict_clip)
         return strict_clip
 
 
@@ -902,7 +902,7 @@ def clip_by_polygon(
     for a in to_list(artist):
         match a:
             case ContourSet() if not _MPL_3_8:
-                artists.extend(a.collections)
+                artists.extend(a.collections)  # type: ignore
             case Artist():
                 artists.append(a)
             case _: