+++ src/frykit/prof.py
@@ -3,12 +3,14 @@ from __future__ import annotations
 import cProfile
 from collections.abc import Callable
 from functools import wraps
-from typing import Any, cast
+from typing import Any, TypeVar, cast
 
 from line_profiler import LineProfiler
 
-from frykit.typing import F, PathType
+from frykit.typing import PathType
 from frykit.utils import deprecator
+
+F = TypeVar("F", bound=Callable)
 
 
 def cprofiler(filepath: PathType) -> Callable[[F], F]: